fix: path traversal

+++ src/main/java/org/owasp/webgoat/webwolf/FileServer.java
@@ -12,6 +12,7 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.attribute.FileTime;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
@@ -73,6 +74,11 @@ public class FileServer {
     // https://stackoverflow.com/questions/60336929/java-nio-file-nosuchfileexception-when-file-transferto-is-called
     try (InputStream is = multipartFile.getInputStream()) {
       var destinationFile = destinationDir.toPath().resolve(multipartFile.getOriginalFilename());
+      Path normalizedPath = destinationFile.normalize();
+      if (!normalizedPath.startsWith(destinationDir.toPath())) {
+        throw new IOException("Invalid file path");
+      }
+ 
       Files.deleteIfExists(destinationFile);
       Files.copy(is, destinationFile);
     }
